Update unidouble program id
It was necessary because I could not upgrade anymore. I could not
upgrade anymore because there were not enough space left on the
account.

To deploy with another program id, I deleted target/deploy/*
then did `anchor deploy`. I run `anchor test` after that and updated
the store creator and program id everywhere in the code.